YQL-19747 Refactor syntax module
